// src/utils/types/ControlType.tsx
export enum ControlType {
    IFRAME = "iframe",
    SUBGRID = "subgrid",
    WEBRESSOURCE = "webresource",
    NOTES = "notes",
    TIMER = "timercontrol",
    KBSEARCH = "kbsearch",
    TIMELINE = "timelinewall",
    QUICKFORM = "quickform",
    LOOKUP = 'lookup',
    STANDARD = 'standard',
    OPTIONSET = 'optionset',
} 